v4/options: split option 121 at route boundary
When there are so many routes the option needs to be split in multiple
options, perform the split at a route boundary instead of an arbitrary
byte boundary at 255.

The RFCs are clear: we can split at an arbitrary byte, and decoders MUST
handle splits in the middle of a logical unit. This is the case for most
DHCP clients but not for systemd. This impacts a few distributions like
Debian 12, Ubuntu 20.04, Ubuntu 22.04, Ubuntu 24.04, etc.

So while we wait for an implementation of RFC3396 in systemd-networkd,
let's just split the option I care about: classless static routes.